Move update handlers around a bit to improve project structure, and fix fw_sort_key bug.
