+++ goosebit/api/firmware.py
@@ -5,7 +5,7 @@ from goosebit.auth import validate_user_permissions
 from goosebit.permissions import Permissions
 from goosebit.settings import UPDATES_DIR
 from goosebit.updater.misc import fw_sort_key
-from goosebit.updater.updates import FirmwareArtifact
+from goosebit.updates.artifacts import FirmwareArtifact
 
 router = APIRouter(prefix="/firmware")
