+++ goosebit/updates/version.py
@@ -2,7 +2,6 @@ import dataclasses
 from datetime import datetime
 from pathlib import Path
 from typing import Callable, Literal
-import time
 
 
 @dataclasses.dataclass
